Fixed bugs
Changed the length of the string for the current score. The previous string
would extend to the next row for two digits. Next, Added an extra expression
for updating the score. Previously, it was checking when the player's row was
GREATER than the enemy row, instead of less than. Also, handles the case where
the player's row and enemy's row are the same. To handle it, I check when both rows
are the same and the player's column is equal to the open column of the enemy. Lastly,
also added DDRC in main, otherwise, the LCD functions would not write to the LCD display.